Extract MeetingListView from the main window
